Remove splitting from mpc-ot and mpc-ot-core
- remove unused state variables
- adapt verify function
- remove barrier
- remove factory implementation
- some OT traits related to garbling have been moved to
  mpc-garble
