+++ mpc/garble/mpc-garble/src/lib.rs
@@ -87,3 +87,66 @@ pub trait Compressor {
         circ: GarbledCircuit<gc_state::Evaluated>,
     ) -> Result<GarbledCircuit<gc_state::Compressed>, GCError>;
 }
+
+#[async_trait]
+impl<T> ObliviousSend<FullEncodedInput> for T
+where
+    T: Send + ObliviousSend<[Block; 2]>,
+{
+    async fn send(&mut self, inputs: Vec<FullEncodedInput>) -> Result<(), OTError> {
+        self.send(
+            inputs
+                .into_iter()
+                .map(|labels| labels.iter_blocks().collect::<Vec<[Block; 2]>>())
+                .flatten()
+                .collect::<Vec<[Block; 2]>>(),
+        )
+        .await
+    }
+}
+
+#[async_trait]
+impl<T> ObliviousReceive<InputValue, ActiveEncodedInput> for T
+where
+    T: Send + ObliviousReceive<bool, Block>,
+{
+    async fn receive(
+        &mut self,
+        choices: Vec<InputValue>,
+    ) -> Result<Vec<ActiveEncodedInput>, OTError> {
+        let choice_bits = choices
+            .iter()
+            .map(|value| value.value().to_bits(value.bit_order()))
+            .flatten()
+            .collect::<Vec<bool>>();
+
+        let mut blocks = self.receive(choice_bits).await?;
+
+        Ok(choices
+            .into_iter()
+            .map(|value| {
+                let labels = ActiveLabels::from_blocks(blocks.drain(..value.len()).collect());
+                ActiveEncodedInput::from_active_labels(value.group().clone(), labels)
+                    .expect("Input labels should be valid")
+            })
+            .collect())
+    }
+}
+
+#[async_trait]
+impl<T> ObliviousVerify<FullEncodedInput> for T
+where
+    T: Send + ObliviousVerify<[Block; 2]>,
+{
+    async fn verify(self, input: Vec<FullEncodedInput>) -> Result<(), OTError> {
+        self.verify(
+            input
+                .into_iter()
+                .map(|labels| labels.iter_blocks().collect::<Vec<[Block; 2]>>())
+                .flatten()
+                .collect(),
+        )
+        .await
+    }
+}
+
