+++ mpc/ot/mpc-ot-core/src/extension/kos15/receiver/error.rs
@@ -15,8 +15,6 @@ pub enum ExtReceiverCoreError {
     InvalidChoiceLength,
     #[error("Received payload of unexpected size")]
     InvalidPayloadSize,
-    #[error("Unable to split OT after derandomization")]
-    SplitAfterDerand,
     #[error("Matrix Error: {0}")]
     MatrixError(#[from] MatrixError),
 }
