+++ mpc/ot/mpc-ot/src/kos/receiver.rs
@@ -74,30 +74,6 @@ impl Kos15IOReceiver<r_state::RandSetup> {
     pub fn remaining(&self) -> usize {
         self.inner.remaining()
     }
-
-    /// Splits OT into separate instances, returning the original instance and the new instance
-    /// respectively.
-    ///
-    /// * channel - Channel to attach to the new instance
-    /// * count - Number of OTs to allocate to the new instance
-    pub fn split(self, channel: OTChannel, split_at: usize) -> Result<(Self, Self), OTError> {
-        let Self {
-            inner: mut child,
-            channel: parent_channel,
-        } = self;
-
-        let parent = Self {
-            inner: child.split(split_at)?,
-            channel: parent_channel,
-        };
-
-        let child = Self {
-            inner: child,
-            channel,
-        };
-
-        Ok((parent, child))
-    }
 }
 
 #[async_trait]
