+++ mpc/ot/mpc-ot/src/lib.rs
@@ -4,14 +4,11 @@ pub mod mock;
 
 use async_trait::async_trait;
 use futures::channel::oneshot::Canceled;
-use mpc_circuits::{InputValue, WireGroup};
-use mpc_core::Block;
-use mpc_garble_core::{ActiveEncodedInput, ActiveLabels, FullEncodedInput};
 use mpc_ot_core::{
-    msgs::{OTFactoryMessage, OTMessage},
-    CommittedOTError, ExtReceiverCoreError, ExtSenderCoreError, ReceiverCoreError, SenderCoreError,
+    msgs::OTMessage, CommittedOTError, ExtReceiverCoreError, ExtSenderCoreError, ReceiverCoreError,
+    SenderCoreError,
 };
-use utils_aio::{mux::MuxerError, Channel};
+use utils_aio::Channel;
 
 pub use mpc_ot_core::config;
 
@@ -37,22 +34,6 @@ pub enum OTError {
     InvalidCiphertextLength(usize, usize),
     #[error("Encountered error in backend")]
     Backend(#[from] Canceled),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum OTFactoryError {
-    #[error("muxer error")]
-    MuxerError(#[from] MuxerError),
-    #[error("ot error")]
-    OTError(#[from] OTError),
-    #[error("io error")]
-    IOError(#[from] std::io::Error),
-    #[error("unexpected message")]
-    UnexpectedMessage(OTFactoryMessage),
-    #[error("{0} Sender expects {1} OTs, Receiver expects {2}")]
-    SplitMismatch(String, usize, usize),
-    #[error("other: {0}")]
-    Other(String),
 }
 
 #[async_trait]
@@ -89,71 +70,8 @@ pub trait ObliviousVerify<T> {
     async fn verify(self, input: Vec<T>) -> Result<(), OTError>;
 }
 
-#[async_trait]
-impl<T> ObliviousSend<FullEncodedInput> for T
-where
-    T: Send + ObliviousSend<[Block; 2]>,
-{
-    async fn send(&mut self, inputs: Vec<FullEncodedInput>) -> Result<(), OTError> {
-        self.send(
-            inputs
-                .into_iter()
-                .map(|labels| labels.iter_blocks().collect::<Vec<[Block; 2]>>())
-                .flatten()
-                .collect::<Vec<[Block; 2]>>(),
-        )
-        .await
-    }
-}
-
-#[async_trait]
-impl<T> ObliviousReceive<InputValue, ActiveEncodedInput> for T
-where
-    T: Send + ObliviousReceive<bool, Block>,
-{
-    async fn receive(
-        &mut self,
-        choices: Vec<InputValue>,
-    ) -> Result<Vec<ActiveEncodedInput>, OTError> {
-        let choice_bits = choices
-            .iter()
-            .map(|value| value.value().to_bits(value.bit_order()))
-            .flatten()
-            .collect::<Vec<bool>>();
-
-        let mut blocks = self.receive(choice_bits).await?;
-
-        Ok(choices
-            .into_iter()
-            .map(|value| {
-                let labels = ActiveLabels::from_blocks(blocks.drain(..value.len()).collect());
-                ActiveEncodedInput::from_active_labels(value.group().clone(), labels)
-                    .expect("Input labels should be valid")
-            })
-            .collect())
-    }
-}
-
-#[async_trait]
-impl<T> ObliviousVerify<FullEncodedInput> for T
-where
-    T: Send + ObliviousVerify<[Block; 2]>,
-{
-    async fn verify(self, input: Vec<FullEncodedInput>) -> Result<(), OTError> {
-        self.verify(
-            input
-                .into_iter()
-                .map(|labels| labels.iter_blocks().collect::<Vec<[Block; 2]>>())
-                .flatten()
-                .collect(),
-        )
-        .await
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::mock::mock_ot_pair;
     use mpc_circuits::ADDER_64;
     use mpc_garble_core::FullInputSet;
